+++ upcloud/client/client.go
@@ -15,7 +15,7 @@ import (
 )
 
 const (
-	Version    string = "8.4.0"
+	Version    string = "8.5.0"
 	APIVersion string = "1.3"
 	APIBaseURL string = "https://api.upcloud.com"